% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wtsmth_Fit.R
\name{fit_WTSMTH}
\alias{fit_WTSMTH}
\title{Need A Title For This Function}
\usage{
fit_WTSMTH(
  data,
  lambda1,
  lambda2,
  family = c("gaussian", "binomial"),
  iter.control = list(max_iter = 8L, tol_beta = 10^(-3), tol_loss = 10^(-6)),
  ...
)
}
\arguments{
\item{data}{An object of class "WTsth.data" as generated by prep()}

\item{lambda1}{A numeric vector. Lambda_1 values to be considered. Provided
values will be transformed to 2^(lambda1).}

\item{lambda2}{A numeric vector Lambda_2 values to be considered. Provided
values will be transformed to 2^(lambda2).}

\item{family}{A character. The family of the outcome. Must be one of
"gaussian" (Y is continuous) or "binomial" (Y is binary).}

\item{iter.control}{A list object. Allows user to control iterative
update procedure. Allowed elements are "max.iter", the maximum number
of iterations; "tol.beta", the difference between consecutive beta
updates below which the procedure is deemed converged; and "tol.loss",
the difference in consecutive loss updates below which the procedure
is deemed converged.}

\item{cv.control}{A list object. Allows user to control cross-validation
procedure. Allowed elements are "n.fold", the number of cross-validation
folds; "n.core", the number of cores to use in procedure; and
"stratified", if TRUE and family = "binomial", the folds will be
stratified (this option should be used if either category is
"rare.")}
}
\value{
A numeric vector. The estimated model parameters
}
\description{
Performs weighted smooth regression for a given lambda1, lambda2 combination.
}
\examples{

# we need to create a data set and add an example

}
